Test commit

+++ src/main/java/com/rbkmoney/fraudo/p2p/resolver/P2PTimeWindowResolver.java
@@ -14,7 +14,7 @@ public class P2PTimeWindowResolver implements TimeWindowResolver<Time_windowCont
     @Override
     public TimeWindow resolve(Time_windowContext ctx) {
         TimeWindow.TimeWindowBuilder builder = TimeWindow.builder();
-        List<TerminalNode> times = ctx.DECIMAL();
+        List<TerminalNode> times = ctx.INTEGER();
         String startWindow = TextUtil.safeGetText(times.get(0));
         if (times.size() == 2) {
             String endWindow = TextUtil.safeGetText(times.get(1));
